requirements.txt
# requirements.txt for BIDS-ETL
flask>=2.0
pandas>=1.3
elasticsearch>=8.0
elasticsearch-helpers>=8.0
bids-layout>=0.17
python-dotenv>=0.19
requests>=2.26
tqdm>=4.62
pyyaml>=6.0
